support SAML attributes that control the session expiration
Either SessionDuration or SessionNotOnOrAfter can be provided by the
SAML IdP, and can be defined for up to 12 hours. The shortest expiration
out of those two wins.

If the IdP sends either of these, and the smallest of the two is larger
than 60 minutes, we shouldn't send our own DurationSeconds parameter
when requesting temporary credentials. DurationSeconds may only go up to
60 minutes, so it will actually force the session expiration to be
shorter if we send it at all when either SAML attribute exists and is
bigger than DurationSeconds.

If either of the SAML attributes is smaller than 60 minutes, it will win
regardless of what we set or don't set DurationSeconds to.